Move styleSheetObserver
